% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{plot.cow}
\alias{plot.cow}
\title{Plot method for a object of class 'cow'}
\usage{
\method{plot}{cow}(cow, models, choice = "barplot", label_angle, fixed_value,
  R = 100, ...)
}
\arguments{
\item{cow}{a object of class 'cow' which will be plot}

\item{models}{list of all the models describing the cow attributes}

\item{choice}{choice of the graph among 'barplot' (by default) and 'radar_diag'}

\item{label_angle}{integer between 0 and 360 which is the angle of the x labels for the barplot}

\item{fixed_value}{vector containing the postion of the cow attributes which do not need prediction interval}

\item{R}{integer for the number of predicted value by under-models for prediction interval (100 by default).
Warning, if the number of the under-model is lower than R, the number of under-model will be the max.}
}
\description{
Plot the cow attribute either a barplot with prediction interval, either a radar diagram
}
\examples{
library(mfe)
load(models.RData) #need models fitted
data(indicateurs)
cow0 <- setNames(data.frame(matrix(0, ncol = ncol(indicateurs), nrow = 1)), colnames(indicateurs))
pred_cow <- predict_cow(cow_simulated = cow0, models = models, list_index = 1:3, list_value = c(0, 1, -0.5), B=5, R=10)
plot(pred_cow)

}
\seealso{
\code{\link{predict_cow}}
}
