% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{predict.rf}
\alias{predict.rf}
\title{Predict method for rf model fits}
\usage{
\method{predict}{rf}(model, newdata, predict.all = FALSE, R = 100,
  interval = FALSE, interval_method = "sd", ...)
}
\arguments{
\item{model}{'rf' class object}

\item{newdata}{matrix containg the regressor of the model that you want to predict the response for.}

\item{predict.all}{logical value (FALSE by default).
If TRUE, a matrix of dimension \code{R} by \code{nrow(neawdata)} is return,
containing several predictions from undermodels generate by \code{\link{underModel.rf}}}

\item{interval}{a boolean value (FALSE by default) to return a interval of prediction}

\item{interval_method}{either 'sd' (by default) or 'qt'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predicted values based on rf model object.
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- rf(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
predict(model, newdata = indicateurs[1,])
predict(model, newdata = indicateurs[1,], predict.all = TRUE, R = 10)

}
\seealso{
\code{\link{rf}}, \code{\link{underModel.rf}}
}
