% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{predict.ridge}
\alias{predict.ridge}
\title{Predict method for ridge model fits}
\usage{
\method{predict}{ridge}(model, newdata, predict.all = FALSE, R = 100, ...)
}
\arguments{
\item{model}{'ridge' class object}

\item{newdata}{matrix containg the regressor of the model that you want to predict the response for.}

\item{predict.all}{logical value (FALSE by default).
If TRUE, a matrix of dimension \code{R} by \code{nrow(neawdata)} is return,
containing several predictions from undermodels generate by \code{\link{underModel.ridge}}}

\item{R}{numeric value which set the number of under-prediction made when predict.all=TRUE}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predicted values based on ridge model object.
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- ridge(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
predict(model, newdata = indicateurs[1,])
predict(model, newdata = indicateurs[1,], predict.all = TRUE, R = 10)

}
\seealso{
\code{\link{ridge}}, \code{\link{underModel.ridge}}
}
