% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{predict_cow}
\alias{predict_cow}
\title{Predict a cow}
\usage{
predict_cow(cow_simulated, models, list_index, list_value, R = 10, B = 5)
}
\arguments{
\item{cow_simulated}{a matrix or data.frame 1 by p attributes with the colnames matching with the models}

\item{models}{list of all the models describing the cow attributes}

\item{list_index}{vector containing the position in the data.frame of the fixed value}

\item{list_value}{vector containing the value at the list_index position}

\item{R}{integer to converge toward a stable result (10 by default)}

\item{B}{integer for the number of bootstrap which switch the order of prediction (5 by default)}
}
\value{
a predicted cow througth a object cow
}
\description{
Predict the cow performances with the whole models
}
\examples{
library(mfe)
load(models.RData) #need models fitted
data(indicateurs)
cow0 <- setNames(data.frame(matrix(0, ncol = ncol(indicateurs), nrow = 1)), colnames(indicateurs))
pred_cow <- predict_cow(cow_simulated = cow0, models = models, list_index = 1:3, list_value = c(0, 1, -0.5), B=5, R=10)
plot(pred_cow)


}
\seealso{
\code{\link{plot.cow}}
}
