% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{rf models}
\usage{
rf(Y, X, ntree = 500, mtry, Ylabel = NULL)
}
\arguments{
\item{Y}{numerical vector of size n containing the response to predict.}

\item{X}{numerical matrix of dimension n by p regressors}

\item{Ylabel}{a character value containing the name of the Y varaible}
}
\description{
rf is a method used to fit random forest models using the randomForest package.
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- rf(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
predict(model, newdata = indicateurs[1,])


}
\seealso{
\code{\link{predict.rf}}
}
