% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge models}
\usage{
ridge(X, Y, Ylabel = NULL)
}
\arguments{
\item{X}{numerical matrix of dimension n by p regressors}

\item{Y}{numerical vector of size n containing the response to predict.}

\item{Ylabel}{a character value containing the name of the Y varaible}
}
\description{
Ridge is a method used to fit linear models with constraint lambda using the glmnet package.
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- ridge(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
predict(model, newdata = indicateurs[1,])


}
\seealso{
\code{\link{predict.ridge}}
}
