% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\title{sir models}
\usage{
sir(Y, X, H = 10, K = 1, Ylabel = NULL)
}
\arguments{
\item{Y}{numerical vector of size n containing the response to predict.}

\item{X}{numerical matrix of dimension n by p regressors}

\item{H}{integer (by default 10) which set the number of slices}

\item{K}{integer (by default 1) which set the number of component taken. Only 1 is possible right now.}

\item{Ylabel}{a character value containing the name of the Y varaible}
}
\value{
An object with S3 class "sir" with the following components:

\item{beta}{ }

\item{index}{ }

\item{H}{H parameter passed througth the function}

\item{K}{K parameter passed througth the function }

\item{hopt}{bandwidth value of the kernel tuned to smooth the points and predict new values.
For more explaination see \code{\link{cv_bandwidth}}}

\item{Xlabels}{a character vector containing the p regessor names}

\item{Ylabel}{a character value containing the response variable name}

\item{x_train}{the training matrix X of dimension \code{p} times \code{n}}

\item{y_train}{the training response vector Y of length \code{n}}
}
\description{
sir is a method used to fit sliced-inverse regression models
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- sir(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
predict(model, newdata = indicateurs[1,])


}
\seealso{
\code{\link{predict.sir}}
}
