% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{underModels.reg_lm}
\alias{underModels.reg_lm}
\title{Predict method for reg_lm model fits}
\usage{
underModels.reg_lm(model, B = 100, ...)
}
\arguments{
\item{model}{'reg_lm' class object}

\item{...}{further arguments passed to or from other methods.}

\item{newdata}{matrix containg the regressor of the model that you want to predict the response for.}

\item{predict.all}{logical value (FALSE by default).
If TRUE, a matrix of dimension \code{R} by \code{nrow(neawdata)} is return,
containing several predictions from undermodels generate by \code{\link{underModel.rf}}}

\item{R}{numeric value which set the number of under-prediction made when predict.all=TRUE}
}
\description{
Predicted values based on reg_lm model object.
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- reg_lm(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
predict(model, newdata = indicateurs[1,])
predict(model, newdata = indicateurs[1,], predict.all = TRUE, R = 10)

}
\seealso{
\code{\link{reg_lm}}, \code{\link{underModel.reg_lm}}
}
