% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{underModels.rf}
\alias{underModels.rf}
\title{Generate undermodels to 'rf' class object}
\usage{
underModels.rf(model, B = 100, ...)
}
\arguments{
\item{model}{'rf' class object}

\item{B}{numeric value which is the number of undermodels generate}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list of 'rf' models
}
\description{
Allow to generate undermodels by bootstraping the dataset used for training.
}
\examples{
library(mfe)
data(indicateurs)
X <- indicateurs[, -c(1,2,3)]
Y <- indicateurs[,1]
model <- rf(X = X, Y = Y, Ylabel = colnames(indicateurs)[1])
model$undermodels <- undermodels.rf(model, B = 10)

}
\seealso{
\code{\link{rf}}
}
